requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.4
django-braces==1.14.0
django-environ==0.4.5
django-heroku==0.3.1
django-oauth-toolkit==1.3.3
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.2
environ==1.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
pycparser==2.20
Pygments==2.7.3
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
urllib3==1.26.2


django-filter~=2.4.0